+++ src/test/test_shader.hpp
@@ -26,6 +26,6 @@ class TestShader : public Test {
   void OnRender() override;
 
   void OnImGuiRender() override;
-  void OnViewportChange(size_t width, size_t height) override;
+  void OnViewportChange(uint32_t width, uint32_t height) override;
 };
 }
